Camera Manager

+++ SluaghEngine/includes/Core/CameraManager.h
@@ -15,8 +15,15 @@ namespace SE
 			CameraManager(Graphics::IRenderer* renderer, const EntityManager& entityManager, TransformManager* transformManager);
 			~CameraManager();
 
+
+			/**
+			* @brief	Create a camera for the given entity.
+			*/
 			void Bind(const Entity& entity, float fov = 1.570796f, float aspectRatio = (800.0f/640.0f), float nearP = 0.01f, float farP = 100.0f, const DirectX::XMFLOAT3& pos = { 0.0f,0.0f,0.0f }, const DirectX::XMFLOAT3& rotation = { 0.0f,0.0f,0.0f });
 
+			/**
+			* @brief	Set the camera as the active camera.
+			*/
 			void SetActive(const Entity& entity);
 
 			/**
@@ -48,12 +55,12 @@ namespace SE
 
 			struct CameraData
 			{
-				static const size_t size = sizeof(Entity) + sizeof(uint8_t) + sizeof(float)*4;
+				static const size_t size = sizeof(Entity) + sizeof(size_t) + sizeof(float)*4;
 				size_t allocated = 0;
 				size_t used = 0;
 				void* data = nullptr;
 				Entity* entity;
-				uint8_t* dirty;
+				size_t* dirty;
 				float* fov;
 				float* aspectRatio;
 				float* nearPlane;
